+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/optimizer/LocalPhysicalPlanOptimizerTests.java
@@ -9,7 +9,6 @@ package org.elasticsearch.xpack.esql.optimizer;
 
 import com.carrotsearch.randomizedtesting.annotations.ParametersFactory;
 
-import org.elasticsearch.common.lucene.BytesRefs;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.core.Tuple;
 import org.elasticsearch.index.query.QueryBuilder;
@@ -26,23 +25,17 @@ import org.elasticsearch.xpack.esql.enrich.ResolvedEnrichPolicy;
 import org.elasticsearch.xpack.esql.expression.function.EsqlFunctionRegistry;
 import org.elasticsearch.xpack.esql.parser.EsqlParser;
 import org.elasticsearch.xpack.esql.plan.logical.Enrich;
-import org.elasticsearch.xpack.esql.plan.logical.Eval;
-import org.elasticsearch.xpack.esql.plan.logical.TopN;
 import org.elasticsearch.xpack.esql.plan.physical.AggregateExec;
-import org.elasticsearch.xpack.esql.plan.physical.EnrichExec;
 import org.elasticsearch.xpack.esql.plan.physical.EsQueryExec;
 import org.elasticsearch.xpack.esql.plan.physical.EsStatsQueryExec;
 import org.elasticsearch.xpack.esql.plan.physical.EsStatsQueryExec.Stat;
 import org.elasticsearch.xpack.esql.plan.physical.EstimatesRowSize;
-import org.elasticsearch.xpack.esql.plan.physical.EvalExec;
 import org.elasticsearch.xpack.esql.plan.physical.ExchangeExec;
 import org.elasticsearch.xpack.esql.plan.physical.FieldExtractExec;
-import org.elasticsearch.xpack.esql.plan.physical.FragmentExec;
 import org.elasticsearch.xpack.esql.plan.physical.LimitExec;
 import org.elasticsearch.xpack.esql.plan.physical.LocalSourceExec;
 import org.elasticsearch.xpack.esql.plan.physical.PhysicalPlan;
 import org.elasticsearch.xpack.esql.plan.physical.ProjectExec;
-import org.elasticsearch.xpack.esql.plan.physical.TopNExec;
 import org.elasticsearch.xpack.esql.planner.FilterTests;
 import org.elasticsearch.xpack.esql.planner.Mapper;
 import org.elasticsearch.xpack.esql.planner.PlannerUtils;
@@ -58,9 +51,6 @@ import org.elasticsearch.xpack.ql.expression.ReferenceAttribute;
 import org.elasticsearch.xpack.ql.expression.function.FunctionRegistry;
 import org.elasticsearch.xpack.ql.index.EsIndex;
 import org.elasticsearch.xpack.ql.index.IndexResolution;
-import org.elasticsearch.xpack.ql.plan.logical.Aggregate;
-import org.elasticsearch.xpack.ql.plan.logical.EsRelation;
-import org.elasticsearch.xpack.ql.plan.logical.Limit;
 import org.elasticsearch.xpack.ql.tree.Source;
 import org.elasticsearch.xpack.ql.type.DataTypes;
 import org.elasticsearch.xpack.ql.type.EsField;
@@ -75,10 +65,8 @@ import static org.elasticsearch.xpack.esql.EsqlTestUtils.configuration;
 import static org.elasticsearch.xpack.esql.EsqlTestUtils.loadMapping;
 import static org.elasticsearch.xpack.esql.EsqlTestUtils.withDefaultLimitWarning;
 import static org.elasticsearch.xpack.esql.plan.physical.AggregateExec.Mode.FINAL;
-import static org.elasticsearch.xpack.esql.plan.physical.AggregateExec.Mode.PARTIAL;
 import static org.elasticsearch.xpack.esql.plan.physical.EsStatsQueryExec.StatsType;
 import static org.hamcrest.Matchers.contains;
-import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.is;
@@ -96,7 +84,6 @@ public class LocalPhysicalPlanOptimizerTests extends ESTestCase {
 
     private EsqlParser parser;
     private Analyzer analyzer;
-
     private LogicalPlanOptimizer logicalOptimizer;
     private PhysicalPlanOptimizer physicalPlanOptimizer;
     private Mapper mapper;
@@ -151,47 +138,17 @@ public class LocalPhysicalPlanOptimizerTests extends ESTestCase {
         mapper = new Mapper(functionRegistry);
         EnrichResolution enrichResolution = new EnrichResolution();
         enrichResolution.addResolvedPolicy(
-            "departments",
+            "foo",
             Enrich.Mode.ANY,
             new ResolvedEnrichPolicy(
-                "employee_id",
+                "fld",
                 EnrichPolicy.MATCH_TYPE,
-                List.of("department"),
-                Map.of("", ".enrich-departments-1", "cluster_1", ".enrich-departments-2"),
-                Map.of("department", new EsField("department", DataTypes.KEYWORD, Map.of(), true))
-            )
-        );
-        enrichResolution.addResolvedPolicy(
-            "departments",
-            Enrich.Mode.COORDINATOR,
-            new ResolvedEnrichPolicy(
-                "employee_id",
-                EnrichPolicy.MATCH_TYPE,
-                List.of("department"),
-                Map.of("", ".enrich-departments-3"),
-                Map.of("department", new EsField("department", DataTypes.KEYWORD, Map.of(), true))
-            )
-        );
-        enrichResolution.addResolvedPolicy(
-            "supervisors",
-            Enrich.Mode.ANY,
-            new ResolvedEnrichPolicy(
-                "department",
-                EnrichPolicy.MATCH_TYPE,
-                List.of("supervisor"),
-                Map.of("", ".enrich-supervisors-a", "cluster_1", ".enrich-supervisors-b"),
-                Map.of("supervisor", new EsField("supervisor", DataTypes.KEYWORD, Map.of(), true))
-            )
-        );
-        enrichResolution.addResolvedPolicy(
-            "supervisors",
-            Enrich.Mode.COORDINATOR,
-            new ResolvedEnrichPolicy(
-                "department",
-                EnrichPolicy.MATCH_TYPE,
-                List.of("supervisor"),
-                Map.of("", ".enrich-supervisors-c"),
-                Map.of("supervisor", new EsField("supervisor", DataTypes.KEYWORD, Map.of(), true))
+                List.of("a", "b"),
+                Map.of("", "idx"),
+                Map.ofEntries(
+                    Map.entry("a", new EsField("a", DataTypes.INTEGER, Map.of(), true)),
+                    Map.entry("b", new EsField("b", DataTypes.LONG, Map.of(), true))
+                )
             )
         );
         analyzer = new Analyzer(
@@ -507,261 +464,6 @@ public class LocalPhysicalPlanOptimizerTests extends ESTestCase {
         var field = as(project.child(), FieldExtractExec.class);
         var fields = field.attributesToExtract();
         assertThat(Expressions.names(fields), contains("_meta_field", "gender", "job", "job.raw", "languages", "long_noidx"));
-    }
-
-    public void testEnrichBeforeAggregation() {
-        {
-            var plan = physicalPlan("""
-                from test
-                | eval employee_id = to_str(emp_no)
-                | ENRICH[ccq.mode:any] departments
-                | STATS size=count(*) BY department""");
-            var limit = as(plan, LimitExec.class);
-            var finalAggs = as(limit.child(), AggregateExec.class);
-            assertThat(finalAggs.getMode(), equalTo(FINAL));
-            var exchange = as(finalAggs.child(), ExchangeExec.class);
-            var fragment = as(exchange.child(), FragmentExec.class);
-            var partialAggs = as(fragment.fragment(), Aggregate.class);
-            var enrich = as(partialAggs.child(), Enrich.class);
-            assertThat(enrich.mode(), equalTo(Enrich.Mode.ANY));
-            assertThat(enrich.concreteIndices(), equalTo(Map.of("", ".enrich-departments-1", "cluster_1", ".enrich-departments-2")));
-            var eval = as(enrich.child(), Eval.class);
-            as(eval.child(), EsRelation.class);
-        }
-        {
-            var plan = physicalPlan("""
-                from test
-                | eval employee_id = to_str(emp_no)
-                | ENRICH[ccq.mode:coordinator] departments
-                | STATS size=count(*) BY department""");
-            var limit = as(plan, LimitExec.class);
-            var finalAggs = as(limit.child(), AggregateExec.class);
-            assertThat(finalAggs.getMode(), equalTo(FINAL));
-            var partialAggs = as(finalAggs.child(), AggregateExec.class);
-            assertThat(partialAggs.getMode(), equalTo(PARTIAL));
-            var enrich = as(partialAggs.child(), EnrichExec.class);
-            assertThat(enrich.mode(), equalTo(Enrich.Mode.COORDINATOR));
-            assertThat(enrich.concreteIndices(), equalTo(Map.of("", ".enrich-departments-3")));
-            var exchange = as(enrich.child(), ExchangeExec.class);
-            var fragment = as(exchange.child(), FragmentExec.class);
-            var eval = as(fragment.fragment(), Eval.class);
-            as(eval.child(), EsRelation.class);
-        }
-    }
-
-    public void testEnrichBeforeLimit() {
-        {
-            var plan = physicalPlan("""
-                from test
-                | eval employee_id = to_str(emp_no)
-                | ENRICH[ccq.mode:any] departments
-                | LIMIT 10""");
-            var enrich = as(plan, EnrichExec.class);
-            assertThat(enrich.mode(), equalTo(Enrich.Mode.ANY));
-            assertThat(enrich.concreteIndices(), equalTo(Map.of("", ".enrich-departments-1", "cluster_1", ".enrich-departments-2")));
-            var eval = as(enrich.child(), EvalExec.class);
-            var finalLimit = as(eval.child(), LimitExec.class);
-            var exchange = as(finalLimit.child(), ExchangeExec.class);
-            var fragment = as(exchange.child(), FragmentExec.class);
-            var partialLimit = as(fragment.fragment(), Limit.class);
-            as(partialLimit.child(), EsRelation.class);
-        }
-        {
-            var plan = physicalPlan("""
-                from test
-                | eval employee_id = to_str(emp_no)
-                | ENRICH[ccq.mode:coordinator] departments
-                | LIMIT 10""");
-            var enrich = as(plan, EnrichExec.class);
-            assertThat(enrich.mode(), equalTo(Enrich.Mode.COORDINATOR));
-            assertThat(enrich.concreteIndices(), equalTo(Map.of("", ".enrich-departments-3")));
-            var eval = as(enrich.child(), EvalExec.class);
-            var finalLimit = as(eval.child(), LimitExec.class);
-            var exchange = as(finalLimit.child(), ExchangeExec.class);
-            var fragment = as(exchange.child(), FragmentExec.class);
-            var partialLimit = as(fragment.fragment(), Limit.class);
-            as(partialLimit.child(), EsRelation.class);
-        }
-    }
-
-    public void testEnrichBeforeTopN() {
-        {
-            var plan = physicalPlan("""
-                from test
-                | eval employee_id = to_str(emp_no)
-                | ENRICH[ccq.mode:any] departments
-                | SORT department
-                | LIMIT 10""");
-            var topN = as(plan, TopNExec.class);
-            var exchange = as(topN.child(), ExchangeExec.class);
-            var fragment = as(exchange.child(), FragmentExec.class);
-            var partialTopN = as(fragment.fragment(), TopN.class);
-            var enrich = as(partialTopN.child(), Enrich.class);
-            assertThat(enrich.mode(), equalTo(Enrich.Mode.ANY));
-            assertThat(enrich.concreteIndices(), equalTo(Map.of("", ".enrich-departments-1", "cluster_1", ".enrich-departments-2")));
-            var eval = as(enrich.child(), Eval.class);
-            as(eval.child(), EsRelation.class);
-        }
-        {
-            var plan = physicalPlan("""
-                from test
-                | eval employee_id = to_str(emp_no)
-                | ENRICH[ccq.mode:coordinator] departments
-                | SORT department
-                | LIMIT 10""");
-            var topN = as(plan, TopNExec.class);
-            var enrich = as(topN.child(), EnrichExec.class);
-            assertThat(enrich.mode(), equalTo(Enrich.Mode.COORDINATOR));
-            assertThat(enrich.concreteIndices(), equalTo(Map.of("", ".enrich-departments-3")));
-            var exchange = as(enrich.child(), ExchangeExec.class);
-            var fragment = as(exchange.child(), FragmentExec.class);
-            var eval = as(fragment.fragment(), Eval.class);
-            as(eval.child(), EsRelation.class);
-        }
-    }
-
-    public void testEnrichAfterTopN() {
-        {
-            var plan = physicalPlan("""
-                from test
-                | SORT emp_no
-                | LIMIT 10
-                | eval employee_id = to_str(emp_no)
-                | ENRICH[ccq.mode:any] departments
-                """);
-            var enrich = as(plan, EnrichExec.class);
-            assertThat(enrich.mode(), equalTo(Enrich.Mode.ANY));
-            assertThat(enrich.concreteIndices(), equalTo(Map.of("", ".enrich-departments-1", "cluster_1", ".enrich-departments-2")));
-            var eval = as(enrich.child(), EvalExec.class);
-            var topN = as(eval.child(), TopNExec.class);
-            var exchange = as(topN.child(), ExchangeExec.class);
-            var fragment = as(exchange.child(), FragmentExec.class);
-            var partialTopN = as(fragment.fragment(), TopN.class);
-            as(partialTopN.child(), EsRelation.class);
-        }
-        {
-            var plan = physicalPlan("""
-                from test
-                | SORT emp_no
-                | LIMIT 10
-                | eval employee_id = to_str(emp_no)
-                | ENRICH[ccq.mode:coordinator] departments
-                """);
-            var enrich = as(plan, EnrichExec.class);
-            assertThat(enrich.mode(), equalTo(Enrich.Mode.COORDINATOR));
-            assertThat(enrich.concreteIndices(), equalTo(Map.of("", ".enrich-departments-3")));
-            var eval = as(enrich.child(), EvalExec.class);
-            var topN = as(eval.child(), TopNExec.class);
-            var exchange = as(topN.child(), ExchangeExec.class);
-            var fragment = as(exchange.child(), FragmentExec.class);
-            var partialTopN = as(fragment.fragment(), TopN.class);
-            as(partialTopN.child(), EsRelation.class);
-        }
-    }
-
-    public void testManyEnrich() {
-        {
-            var plan = physicalPlan("""
-                from test
-                | eval employee_id = to_str(emp_no)
-                | ENRICH[ccq.mode:any] departments
-                | SORT emp_no
-                | LIMIT 100
-                | ENRICH[ccq.mode:any] supervisors
-                | STATS teams=count(*) BY supervisor
-                """);
-            var limit = as(plan, LimitExec.class);
-            var finalAgg = as(limit.child(), AggregateExec.class);
-            var partialAgg = as(finalAgg.child(), AggregateExec.class);
-            var enrich1 = as(partialAgg.child(), EnrichExec.class);
-            assertThat(enrich1.policyName(), equalTo("supervisors"));
-            assertThat(enrich1.mode(), equalTo(Enrich.Mode.ANY));
-            var finalTopN = as(enrich1.child(), TopNExec.class);
-            var exchange = as(finalTopN.child(), ExchangeExec.class);
-            var fragment = as(exchange.child(), FragmentExec.class);
-            var partialTopN = as(fragment.fragment(), TopN.class);
-            var enrich2 = as(partialTopN.child(), Enrich.class);
-            assertThat(BytesRefs.toString(enrich2.policyName().fold()), equalTo("departments"));
-            assertThat(enrich2.mode(), equalTo(Enrich.Mode.ANY));
-            var eval = as(enrich2.child(), Eval.class);
-            as(eval.child(), EsRelation.class);
-        }
-        {
-            var plan = physicalPlan("""
-                from test
-                | eval employee_id = to_str(emp_no)
-                | ENRICH[ccq.mode:any] departments
-                | SORT emp_no
-                | LIMIT 100
-                | ENRICH[ccq.mode:coordinator] supervisors
-                | STATS teams=count(*) BY supervisor
-                """);
-            var limit = as(plan, LimitExec.class);
-            var finalAgg = as(limit.child(), AggregateExec.class);
-            var partialAgg = as(finalAgg.child(), AggregateExec.class);
-            var enrich1 = as(partialAgg.child(), EnrichExec.class);
-            assertThat(enrich1.policyName(), equalTo("supervisors"));
-            assertThat(enrich1.mode(), equalTo(Enrich.Mode.COORDINATOR));
-            var finalTopN = as(enrich1.child(), TopNExec.class);
-            var exchange = as(finalTopN.child(), ExchangeExec.class);
-            var fragment = as(exchange.child(), FragmentExec.class);
-            var partialTopN = as(fragment.fragment(), TopN.class);
-            var enrich2 = as(partialTopN.child(), Enrich.class);
-            assertThat(BytesRefs.toString(enrich2.policyName().fold()), equalTo("departments"));
-            assertThat(enrich2.mode(), equalTo(Enrich.Mode.ANY));
-            var eval = as(enrich2.child(), Eval.class);
-            as(eval.child(), EsRelation.class);
-        }
-        {
-            var plan = physicalPlan("""
-                from test
-                | eval employee_id = to_str(emp_no)
-                | ENRICH[ccq.mode:coordinator] departments
-                | SORT emp_no
-                | LIMIT 100
-                | ENRICH[ccq.mode:any] supervisors
-                | STATS teams=count(*) BY supervisor
-                """);
-            var limit = as(plan, LimitExec.class);
-            var finalAgg = as(limit.child(), AggregateExec.class);
-            var partialAgg = as(finalAgg.child(), AggregateExec.class);
-            var enrich1 = as(partialAgg.child(), EnrichExec.class);
-            assertThat(enrich1.policyName(), equalTo("supervisors"));
-            assertThat(enrich1.mode(), equalTo(Enrich.Mode.ANY));
-            var topN = as(enrich1.child(), TopNExec.class);
-            var enrich2 = as(topN.child(), EnrichExec.class);
-            assertThat(enrich2.policyName(), equalTo("departments"));
-            assertThat(enrich2.mode(), equalTo(Enrich.Mode.COORDINATOR));
-            var exchange = as(enrich2.child(), ExchangeExec.class);
-            var fragment = as(exchange.child(), FragmentExec.class);
-            var eval = as(fragment.fragment(), Eval.class);
-            as(eval.child(), EsRelation.class);
-        }
-        {
-            var plan = physicalPlan("""
-                from test
-                | eval employee_id = to_str(emp_no)
-                | ENRICH[ccq.mode:coordinator] departments
-                | SORT emp_no
-                | LIMIT 100
-                | ENRICH[ccq.mode:any] supervisors
-                | STATS teams=count(*) BY supervisor
-                """);
-            var limit = as(plan, LimitExec.class);
-            var finalAgg = as(limit.child(), AggregateExec.class);
-            var partialAgg = as(finalAgg.child(), AggregateExec.class);
-            var enrich1 = as(partialAgg.child(), EnrichExec.class);
-            assertThat(enrich1.policyName(), equalTo("supervisors"));
-            assertThat(enrich1.mode(), equalTo(Enrich.Mode.ANY));
-            var topN = as(enrich1.child(), TopNExec.class);
-            var enrich2 = as(topN.child(), EnrichExec.class);
-            assertThat(enrich2.policyName(), equalTo("departments"));
-            assertThat(enrich2.mode(), equalTo(Enrich.Mode.COORDINATOR));
-            var exchange = as(enrich2.child(), ExchangeExec.class);
-            var fragment = as(exchange.child(), FragmentExec.class);
-            var eval = as(fragment.fragment(), Eval.class);
-            as(eval.child(), EsRelation.class);
-        }
     }
 
     private QueryBuilder wrapWithSingleQuery(QueryBuilder inner, String fieldName, Source source) {
